Ugrade Tests to MSBUILD `csproj` and Change Pack
Makes the tests runnable with `dotnet test` and updates the packaging
to use the new `dotnet pack` format so that metadata is read directly
from the csproj. This means that a `nuspec` file is no longer needed
and project reference metadata can't get out of sync with the actual
build.

 * Update Tests Project to Compile with Dotnet Build.

   This means that the tests can now be run with `$ dotnet test`.

 * Update the pack to use the `.csproj`. This should reduce the
   duplication of metadata.